requirements.txt
requests
beautifulsoup4
openpyxl
PyYAML
tqdm
lxml
fake-useragent
chardet